analytics/requirements.txt
numpy>=1.24.0 
streamlit
plotly
pandas
python-dotenv
fastapi
uvicorn
requests 